Added CC2420 driver support to ST Discovery board (MCU = STM32F407)
git-svn-id: svn://svn.tuxfamily.org/svnroot/erika/erikae/repos/ee/trunk/ee@2837 88d5ba13-baa4-4915-ae7a-41304cef7dd7

+++ contrib/drivers/radio/cc2420/inc/cc2420.h
@@ -485,7 +485,7 @@ void	cc2420_set_rx(void);
 * @param[in] sleep_level 	The sleep mode: POWER_DOWN or IDLE
 *
  */
-void	cc2420_set_sleep(int);
+void	cc2420_set_sleep(int16_t);
 
 /**
 * @brief Start the radio
